refact: rename info page to about
