--- lib/features/info/presentation/screens/screens.dart
@@ -1,1 +1,0 @@
-export 'info_screen.dart';
